grpc: monitoring: fix method label query
In https://github.com/sourcegraph/sourcegraph/pull/55498, I removed the extra namespace from all the gRPC service metrics.

However, I missed that the method variable (which we use to be able to filter metrics for an individual method) definition was still referencing the old namespace. This had the inadvertent effect of breaking most of the gRPC dashboards on sourcegraph.com - since the metrics they were looking for didn't exist.

This PR fixes this by replacing the namespace in the definition with a grpc_service label filter.