Remove properties from requireRestart that don't actually do (#55499)
This removes a bunch of properties from the list of ones that will show
the site-admin a banner that they need to restart their server, because
they don't actually do (anymore).
This hopefully reduces the number of times an admin has to interrupt
their workflow to go into the infra and restart. This will also benefit
cloud ops where the customer cannot do it themselves.

Reasoning for removal: **Please help me vet that I am not wrong here!!**

- `gitHubApp`: This property is deprecated and not used anymore.
- `codeIntelAutoIndexing.enabled`: This one does not require a server
restart, all the background jobs will be running, but return early. The
next run they will just work as expected. It does require a page
refresh, but that's handled in the SiteAdminConfigurationPage.
- `extensions`: Property no longer exists.
- `auth.sessionExpiry`: This value is used in conf.ContributeValidator
and on the request path for SetActor, it's reading the value dynamically
so this should not require a restart.
- `auth.accessTokens`: Always read dynamically in resolvers/http
handlers. Requires a page refresh, but not a server restart.
- `auth.userOrgMap`: Just used on the user create request path, read
dynamically.
- `update.channel`: Fixed by reading the value on each iteration vs not
starting the ticker.
- `git.cloneURLToRepositoryName`: Uses conf.Cached, should not require a
server restart. None of the accessors of the value seem to cache it
anyhow.
- `searchScopes`: Changelog says this was removed, and it doesn't exist
anymore in the site.schema.json file.

Maybe next time we can get rid of the remaining ones and entirely
eliminate this mechanism for our admins!

## Test plan

Read code carefully to verify that these properties actually don't
require a server restart.