Add database layer for code hosts table (#55402)
Closes https://github.com/sourcegraph/sourcegraph/issues/55435

This adds a new database store to interact with the code_hosts table, as
well as the required hooks into the external_services table so that the
code hosts table always stays up to date with the external_services
table. Later, we will probably want to explicitly create code hosts,
when we redo this entire flow a bit, but for now this will be implicit.
I looked around the code base to see if we modify external_services
elsewhere but that does not seem to be the case. Nice.

The next piece to this will be a OOB migrator that makes it so that all
external_services will eventually have the code_host_id set. In 5.2,
this migration will be introduced then and in 5.3 we will deprecate it,
at which point we can make the code_host_id column non-nullable and
complete this migration.

Still a WIP, need to resolve a few TODO comments and add tests.

Added tests

---------

Co-authored-by: Idan Varsano <varsanojidan@gmail.com>
Co-authored-by: Thorsten Ball <mrnugget@gmail.com>