Begin inferring file type from file content
Go To Definition needs to be able to find the language from the file,
which is currently being done by mapping file name/extension to a single language.
That's not correct, at least part of the time (current case is `.m` files),
but to get the proper language, the file content must be involved.

Change `getModeFromExactFilename` and `getModeFromExtension` to return
string arrays so that the array length can be tested to see if multiple matches
for a file name/extension exist.
Add `getModeFromPathOrContent` to take advantage of that change.
Add `inferModeFromContent` as a placeholder function to be implemented later.
Add lots of comments describing how I think file content inference will work.