sveltekit: Show readme on repo and tree pages
This PR changes the implementation of the repo page to show the contents
of the readme file, if it exist. If not it will show the repository's
description.
The file tree listing was removed. IMO the file tree sidebar works well
suffices.

Tree pages now show a (for now simple) file table, followed by a readme,
if available.