gitserver: Removed test that was never enabled
This test was never enabled since it was added to the code base and upon inspection this test actually clones a repo and stuff, which should not happen in the client tests. I'm removing this for now, if someone feels strongly about re-adding a similar test please go for it.