gitserver: Removed test that was never enabled (#55617)
This test was never enabled since it was first added to the code base
and upon inspection this test actually clones a repo and stuff, which
should not happen in the client tests. I don't see a strong reason for
it to exist, given there are few other tests like this in the clients
package, so I'm removing this for now. If someone feels strongly about
re-adding a similar less overloaded test, please go for it. I also
cleaned up some other conf mocks that I first thought to maybe have been
the culprit.

Closes https://github.com/sourcegraph/sourcegraph/issues/55393

## Test plan

Removal of a test.