httpcli: respect retry-after on all error responses, set retry-after in Cody Gateway responses
This tries to mitigate a few issues we've been seeing in Cody Gateway,
namely that we have 2 layers of retries:

1. Cody Gateway retries upstream
2. Sourcegraph retries Cody Gateway

When we get a series of errors from upstream, e.g. if we get momentarily
rate-limited due to concurrent requests, this instantly causes request
count to explode as Cody Gateway retries, then Sourcegraph retries. Most
of these retries never go anywhere, we just end up blocking more and
more requests as they aggressively retry (our jitter algorithm goes from
0 and 200ms * 2^attempt capped at 3s by default for external services).

This PR implements three changes to try and mitigate this problem on the
Sourcegraph end:

1. `httpcli`: When deciding whether or not to retry, we now always
attempt to respect `retry-after` header on _all_ error responses, since
[503 `retry-after` is part of the standard it seems anyway](
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Retry-After),
when deciding whether a retry is likely to help the situation (i.e. it
must be resolvable within
`SRC_HTTP_CLI_EXTERNAL_RETRY_AFTER_MAX_DURATION=3s`). If no
`retry-after` is available, the old behaviour is preserved (always
retry).
2. `httpcli`: When deciding the _delay_ between retries, instead of
relying only on our jitter algorithm, we now also try and interrogate
`retry-after` to extract a duration to wait while respecting the
configured base (`SRC_HTTP_CLI_EXTERNAL_RETRY_DELAY_BASE`) and maximum
(`SRC_HTTP_CLI_EXTERNAL_RETRY_DELAY_MAX`). This allows us to tell
Sourcegraph clients to not even bother with shorter-delay attempts, and
to just wait for as long as we are allowed to. If the wait is too long,
the retry policy should indicate that. If `retry-after` is not
available, the old behaviour is preserved (exponential jitter delay)
3. Cody Gateway: we now always set `retry-after` from upstream (right
now neither Anthropic or OpenAI provide this, but we do this just in
case they do), otherwise set a default `retry-after`: for Anthropic this
is 2s, since they count concurrents that could be resolved by trying
again on a longer interval, and for OpenAI this is 30s for completions
and longer for embeddings since they use a minute-based rate-limiting
window (see docstrings), so retrying quickly seems unlikely to yield a
positive result

This should help lower the velocity at which we get retries without
changing the defaults or default behaviours in most cases. Additionally,
we're lowering the number of times Cody Gateway attempts to reach its
upstream clients:
sourcegraph/infrastructure#5293

Unit tests on httpcli helpers