Batch Changes - commit message on stdin
Commit messages for batch changes can be arbitrary strings,
so passing them inline to the command arguments using `-m`
is prone to problems and edge cases.
Solving those problems and edge cases is non-optimal,
so pivot to using a different way to specify commit messages:
from a file (stdin in this case). This allows commit messages to contain
arbitrary characters without needing to worry about escaping or edge cases.