gitserver: Start cleaning up server struct
This PR starts moving some things off of the server struct that don't actually need it. What do we gain?
In tests, we don't need to construct the entire server each time, of which we have no clue what of the fields are actually required in a test. This makes the individual functions much easier to test.
As a side effect of some of these things becoming standalone routines, we get:
- Them being goroutines they are tracked in the background routines in our site-admin area
- And I personally find the Main function already much easier to read and reason about