Add site admin banner warning when exceeding Cody limits (#55460)
This PR adds a mechanism that syncs the rate limits in the background,
stores the usage in redis every 30 minutes, and displays a banner
warning when the rate limits are being exceeded:
- Info at 75%+
- Warning at 90%+
- Error at 100%

This PR also moves the code for talking to Cody Gateway to get usage
data into the common codebase out of app, so that we get all the data
exposed through graphql for every instance. This may be useful for
debugging, or implementing additional APIs to visualize limits.

<img width="1728" alt="Screenshot 2023-08-01 at 00 19 49@2x"
src="https://github.com/sourcegraph/sourcegraph/assets/19534377/5e268ed4-af74-414d-bd34-1c299b15c85d">


## Test plan

Verified banners show correctly.