Chore: remove a handful of unused code (#55686)
This removes some unused code as reported by the new [unused
linter](https://github.com/sourcegraph/sourcegraph/pull/55647).

The linter incorrectly reports that some things are unused if they are
only used by tests. It's unclear to me if there is a general solution to
this issue, but for now I'm either ignoring them or moving test-only
stuff to test files.