Fix: Repo embedding scheduler job routine as internal actor
We received reports in slack of repositories never getting scheduled for
embedding jobs despite the filter / pattern preview showing repos that
match.

In both cases linked private repos appear to be impacted. After fetching
repo ids from GetEmbeddableRepos we then make another fetch for getting
the repo names for those repo ids. This read operation is where we
appear to end up missing out on some repos because we are not an
internal actor and therefore we must have permissions to the repos that
we want Listed. When we aren't authorized for any repos then we end up
never scheduling those repos since we never got the repo names.