sveltekit: Remove faker from production build
I assumed that faker wouldn't be included in production builds due to
tree-shaking but I was wrong. It currently adds > 2MB to the production
build files.

I confirmed that the `if` statement is removed in the production
build (i.e. faker is not referenced) but the module is still included.

I thought maybe the dead code removal happens after dependencies are
processed, so I removed the callsite completel but the module would
still be included.

I tried various
[`treeshake`](https://rollupjs.org/configuration-options/#treeshake)
options, such as `moduleSideEffects: false`, but to no avail.

So for the time being, removing the import alltoghether is the best way
to ensure that the module is not included in the production build.