Improve performance of gitserver_repos store methods
We recently added a left join on another table holding the gitserver sync logs, which added quite some overhead to the queries. Here's an example query plan from dotcom for the iterate gitserver repos query:

Before

```
TODO
```

After

```
TODO
```

Since we only used this field in a single place, I decided to create a separate query function for it, in favor of dealing with complexity of conditionally loading or not loading this field.