database: add `repo_update_jobs` table and MVP store.
This is a first incremental change for moving repo-updater in-memory queue into the DB and using workers for cloning/updating repos.

I also have the logic changing the RPC to enqueuing DB jobs in place in other branch, therefore this table is well-tested. The jobs are being scheduled, processed and the results are written back into `last_changed` and `last_fetched` columns.

Test plan:
1) Local sg run with worker code in place and tests that jobs are scheduled and processed. View is used, function works as expected, constraints are enforced (only 1 repo ID enqueued at a certain point of time).
2) sg migration up -> downto -> up.