JetBrains: change plugin name to "Sourcegraph (Cody AI and Code Search)"
The previous "Cody AI by Sourcegraph" name caused confusion for users
who only install this plugin for Code Search functionality. For time
being, the Cody and Code Search functionality is going to live in the
same plugin because it provides the best experience for users of both
products (for example, unified authentication). In the future, we may
end up publishing separate plugins if there's strong enough demand
and we figure out how to keep unified authentication (for example, via
separate "Sourcegraph Authentication" plugin).