blob: Scroll selected line into view even with invalid character offset
When determining whether to scroll a line into the view we are first
validating whether the selected position (which includes line number and
(optionally) character offset) is valid.

It's possible to get a position that refers to a valid line but an
invalid character offset. Currently the selected line is not scrolled
into view in that case.

This change ignores the character offset when validating the position.
The character offset should have no impact on determining whether or not
to scroll the line into view.