search: introduce backend interface for exhaustive search (#55858)
We introduce an interface for how we expect exhaustive search workers will interact with our search subsystem. Additionally we introduce a functional fake for this system which uses a fake search query language to make it easy to test how the parts interact.

Note: The location of the code and names can all change. The purpose of this PR is so we have this code checked in so it can be incrementally improved / easier to track than comments on an issue. I also hope that the fake is useable today to unblock progress.

Test Plan: CI