Implement OOB migration to backfill code_hosts table
This PR implements an out-of-band migration to backfill the code_hosts table with hostnames for existing external services and stores git and api rate limits as well. It is expected to finish in 5.2 on all customer instances, so that we can enforce a NON NULL constraint on the code_host_id column on external_services in 5.3.