doc: add documentation on per-request featureflag overrides
Previously this wasn't documented. The documentation in general is very
much targetted from before we had a UI for this, but just adding this
in.

Ideally I would want to document this better in the client libraries we
use, but the API docs don't really document behaviour and instead rely
on these docs.

Test Plan: n/a