JetBrains: fix `CodyAgent.isConnected` logic
Previously, we would report "agent is not connected" even when we had a
successful agent connection. The root problem was that a string
variable `agentNotRunningExplanation` could be non-empty even when we
were connected. This PR fixes the problem by ignoring the value of
`agentNotRunningExplanation` when we are testing for a connection.