source: implement logic for saving command output and success state of recorded commands (#56109)
Closes [#56044](https://github.com/sourcegraph/sourcegraph/issues/56044)
Closes [#56045](https://github.com/sourcegraph/sourcegraph/issues/56045)

This PR allows Sourcegraph to store a recorded command's success state and output to Redis. This will allow site admins to debug failed commands and understand why they failed adequately.

## Test plan

<!-- All pull requests REQUIRE a test plan: https://docs.sourcegraph.com/dev/background-information/testing_principles 

Why does it matter? 

These test plans are there to demonstrate that are following industry standards which are important or critical for our customers. 
They might be read by customers or an auditor. There are meant be simple and easy to read. Simply explain what you did to ensure 
your changes are correct!

Here are a non exhaustive list of test plan examples to help you:

- Making changes on a given feature or component: 
  - "Covered by existing tests" or "CI" for the shortest possible plan if there is zero ambiguity
  - "Added new tests" 
  - "Manually tested" (if non trivial, share some output, logs, or screenshot)
- Updating docs: 
  - "previewed locally" 
  - share a screenshot if you want to be thorough
- Updating deps, that would typically fail immediately in CI if incorrect
  - "CI" 
  - "locally tested" 
-->

* Set up git recording for any repository on your Sourcegraph instance.
* Access the recorded commands via the redis-cli or any GUI used to access Redis
* There should be new fields `success` and `output` in the data saved for repositories.

![CleanShot 2023-08-22 at 11 33 47@2x](https://github.com/sourcegraph/sourcegraph/assets/25608335/41b1f52c-7ce3-44bf-9d87-5e18663d2afb)