gitserver: Remove repos-stats.json file (#55845)
The content of this file is the size of all repos combined per shard. We update it in the janitor after all repos are processed, but at the same time we collect the same sizes per repo and write them to gitserver_repos.repo_size_bytes. This didn't exist back when the file was introduced, but nowadays, they are pretty much equal. An additional benefit is that other processes that update the repo size in the DB will also reflect faster in the value read here. This removes a bunch of complexity around managing that file, maintaining endpoints to retrieve that data, and around talking to every gitserver to get the cumulated value.

Proof: On dotcom, where 1 gitserver holds roughly 15.5TB of repo data, the calculated value in the file and the value from the DB only differ by 175MB (0.001076729067%).

Please let me know if I am missing some nuance here and those numbers are not actually the same!

With quarterly releases, we can only remove some of the remaining endpoints code in 5.3, to keep compatibility.

There are exactly two things that use this endpoint:

A GraphQL resolver that returns the total git size
If the frontend comes up first, it just uses the DB, which is already populated correctly long ago
If the new gitserver comes up first, it will still have the endpoints but read from the DB, so no problem here
The usagestats code, that collects pings data regularly
If the frontend comes up first, this is not a problem as the DB is already populated correctly and it's read from there
If the gitserver comes up first, it will still have the endpoints but read from the DB, so no problem here