gitserver: Only run disk janitor on Sourcegraph.com (#56202)
This janitor process existed for years and years, and it was originally built for Sourcegraph.com. Its purpose is to make sure that lazily cloned repos don't just take disk space forever.
We can assign dotcom gitservers some disk and it will be able to work with that amount of storage it's given.
On regular instances, however, we are ensuring that ALL repos that the instance knows about are managed by an external service connection, so they are always going to be on disk, and removing them will just trigger an immediate reclone. Thus, there's no good reason to remove them.
We will also get better gitserver disk size alerts with Bolajis current work, so customers will know they have to act, either by removing repos or by increasing disk size.