searchjobs: factor out newSearcher test logic (#56422)
I am busy implementing the NewSearcher for real and this test logic is
useful to reuse. This commit just refactors the existing test case to
make the assertion part re-useable. It also has the nice side-effect of
putting all the bits we assert on next to each other which makes the
test easier to understand.

Test Plan: go test