searcher: always do hybrid search
This makes it so the server side of searcher will never read the
FeatHybrid field in a request, and always treat it as true. This is the
server side step of removing the hybrid search feature flag which has
been rolled out successfully for many many months.

Test Plan: CI