github: fix erratic 404s by keeping URL stable
See [this comment](https://github.com/sourcegraph/customer/issues/2329#issuecomment-1711735942)
and the issue it's in for full context.

Short version:

1. GitHub v3 and v4 clients execute this bit of code multiple times when retrying a request after it ran into rate limits: https://github.com/sourcegraph/sourcegraph/blob/16458d7e5c07ac69d00d32841a1a7dcd55a50ddb/internal/extsvc/github/common.go#L1589-L1590
2. When executed multiple times with the same `url.URL`, `req.URL.Path` ends up with invalid paths like `/api/v3/api/v3`

That lead to 404 errors showing up _sometimes_: when the client runs
into rate limit and retries. And _sometimes_ these 404s were even
ignored and lead to a "successful" end of the syncing process.

This changes the behaviour of the clients to keep a copy of the original
`req.URL` around so that modifications made by `doRequest` don't have
unintended side-effects.