Access tokens: add URL search parameter to control default description
Previously, when using the JetBrains plugin with manual access token
creation, the user had to type out the description of the access token.
This PR adds the ability for the plugin to optionally include a default
description value saving one manual step for the user. With this change,
the user only needs to click "Generate token" instead of
having to additionally type out a description value.