bzl: suppress an irrelevant warning from gazelle
Gazelle wrongly assumes that all files embededed with
embedsrcs are present on disk, while the go_library rule
totally support using targets that generates those inputs.

This has been confusing for teammates. Alas, the only option
we have is to add a gazelle:ignore pragma, so we don't get the
confusing warning.

I added a comment in that BUILD file to reflect that and inform
teammates that if they are to add a new Go file there, they have
to update the rule manually.