repo-updater: Remove repo purge worker
In gitserver, there has been a mechanism in place to wipe repos from disk that have been deleted in the database. It was originally built for sourcegraph.com and never turned on for other instances.
A while ago, we enabled this flag on S2 that makes gitserver delete repos that aren’t present in the database anymore.
Nothing seems to have broken, so I would like to proceed with replacing the repo updater purge worker by only the gitserver janitor process.

What is the repo-updater purge worker?

The repo purge worker in repo updater finds repos that are blocked or deleted in the DB, it will then (after some TTL) call gitserver.RemoveRepo for each of those on an interval.

What is the new process?

The gitserver janitor will look at all repos periodically after each other as part of other cleanup tasks, and query the DB for each repos existence. If it doesn’t exist, is soft-deleted, or blocked, it is deleted from disk, after the same TTL has passed that we respected before. The TTL mechanism here is used to prevent purging all the repos right after some configuration change dropped a bunch of repos.

Known side-effects: The repo purge worker will have missed all repos that are hard deleted from the DB (invalid state, but things can happen), the gitserver purger will see those as well.

Other than that, these processes seem equivalent to me.

Why do that?

- One less thing that repo-updater manages.
- One less thing that potentially removes a repository from disk, making it easier to reason about why a repository has been removed.
  Now, the only ways a repo is deleted are:
  - Manually by a user hitting deleteRepositoryFromDisk through GQL
  - This purge function
  - On dotcom: The additional purge worker that tries to reclaim disk space from repos that are unmanaged
  - By janitor, when corruption is detected
  - By janitor, when a hostname mismatch is detected (after scaling replica count)
  => You can see, all the events that are automated that can trigger a delete now are part of the janitor run! One code path to delete them all :)