Add config to allow for ignoring Perforce protects rules that specify a Host (#56450)
* Change the protects parsing to ignore hosts
Needs to be gated by code host or site config.

* Add a host config setting to control behavior
The setting is "authorization"/"ignoreRulesWithHost" - a boolean value
that defaults to `false`, which preserves current behavior.
When set to `true`, it turns off all processing of protects rules
that have anything other than the wildcard value in the Host field.
Added a test for that behavior in the protects parsing.

* Add CHANGELOG entry

* Adjust tests to accomodate skipping empty lines

* refactor: combine comment and blank line detection

* refactor: change variable/parameter name
`ignoreHostRules` to `ignoreRulesWithHost`
which matches the name of the config field and is a better variable name.

* add docs for "ignoreRulesWithHost"

(cherry picked from commit 7a809147779c502d705cd56a7b84d489e21b4a0c)